Moved SOAP client outside controllers module and added logging functionality for HTTP messages send/retrieval.
