+++ python/nammu/SOAPClient/SOAPClient.py
@@ -3,12 +3,28 @@ import logging
 import httplib as http_client
 import HTTPRequestBuilder
 
+# Dunno if this is the right place to set up logging
+logging.basicConfig()
+
 class SOAPClient:
     """
     Sends and retrieves information to and from the ORACC SOAP server.
     """
     def __init__(self, url):
         self.url = url
+        self.logger, self.request_log = self.set_logger()
+
+    def create_logger(self):
+        """
+        Creates logger to debug HTTP messages sent and responses received.
+        Output should be sent to Nammu's console.
+        """
+        logger = logging.getLogger()
+        logger.setLevel(logging.DEBUG)
+        request_log = logging.getLogger("requests.packages.urllib3")
+        request_log.setLevel(logging.DEBUG)
+        request_log.propagate = True
+        return logger, request_log
 
     def send_message(self):
         """
