+++ python/nammu/test/test_soap.py
@@ -1,6 +1,6 @@
 import pytest
-from ..controller.SOAPClient import SOAPClient
-from ..controller.HTTPRequestBuilder import HTTPRequestBuilder
+from ..SOAPClient.SOAPClient import SOAPClient
+from ..SOAPClient.HTTPRequestBuilder import HTTPRequestBuilder
 
 class TestSOAP:
     def setup(self):
@@ -12,16 +12,19 @@ class TestSOAP:
     def test_send_request(self):
         pass
 
+    def test_http_headers(self):
+        pass
+
+    def test_soap_envelope(self):
+        pass
+
+    def test_http_payload(self):
+        pass
+
     def test_get_response_ok(self):
         pass
 
-    def test_create_soap_envelope(self):
-        pass
-
-    def test_headers(self):
-        pass
-
-    def test_body(self):
+    def test_http_body(self):
         pass
 
     def test_response_ready(self):
